FEM: fem tools package, move fem utils in it
